Move wayland-session / xsession files to a separate target
By moving these to a separate target, they can be easily packaged as
part of a "ui-sway" or "ui-dwm" metapackage. This way we can make the
package manager guarantee that all the dependencies for the ui are
installed when we switch.

The arch packages have done this since the 1.6 release
https://github.com/dreemurrs-embedded/Pine64-Arch/blob/142bd050555414242d8ca325971bb180de01746c/PKGBUILDS/sxmo/sxmo-utils/PKGBUILD#L149

Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>