Fix multikey issues on sway : move to bonsai
To fix the problems mentionned in the following ticket I worked on new
solution to handle complex key bindings on mobile devices:

https://todo.sr.ht/~mil/sxmo-tickets/524

This solution is structured as client/server state machine in hare lang.
It read a json transition tree and wait for the client to send events or
to change contexts.

Installing bonsai is ofc required :

$ doas apk add bonsai

Dont forget to $ sxmo_migrate.sh sync interactive to setup the json in
your config folder.

I adapted the migration scripts to not break the json. We cannot
versionise it so let's just check if the user one differ from the sxmo
default one.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Anjandev Momi <anjan@momi.ca>