enable colour in sway bar and colour some important labels to make them more noticable
swaybar support pango markup, but this has to be explicitly enabled.
Since it offers more customisability for the status line I enabled it
in this patch.

I additionally applied some default colours for the following:

- the status (e.g. LOCK) is now shown in bright red ... This was actually my
  main reason for this implementation as I felt I needed a bit of colour
  to more quickly identify whether the screen was locked or not.
- when battery is low, it will be red (<10%, or orange <20%)
- when modem signal is bad, it will be red
- when battery is charging and is full, it will be green
- the rest remains white

All non-white colours communicate to the user that some action may be
needed.

The implementation is done in such a way that sxmo_status.sh will still
work fine on dwm/ssh (without colours).

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>