no device profile: add warning to ssh and startup
People installing sxmo on new devices often send bug reports with it not
working. Lets improve the error they get when ssh and add a popup on
start referring them to the documentation on how they can fix it.

I also uploaded a commit to sxmo.org with this fix:

https://sxmo.org/deviceprofile

Feel free to change that page's content so it's clearer what the user
should do.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>