appscripts/timer: improved sed call and fixed parsing
Sed is called only once

now it interprets `5m37s` as `5m + 37s` -> `5*60s + 37s` -> `5*60 + 37`
before, it was `5m37s` -> `5*60s37s` -> `5*60*137*1`. the problem is
that the script does not insert a `+` between each non-number characters
and numbers. it does now, and will understand combinations of seconds,
minutes and hours.

Note, the script will interpret `5m37s` and `37s5m` exactly the same way

This may not be necessary if the user uses by themselves`+` to separate
diffent units, though I think the current behavior is very misleading.

Signed-off-by: Anjandev Momi <anjan@momi.ca>