statusbar: remove modem_monitor
There's no reason for this.  I haven't had a dbus-monitor process fail
in, like, forever, and this would not be the point of failure anyway:
they remain even when the main processes go away.

We *do* want some sort of airplane mode, but this isn't that, since this
won't shut down the modem itself or put it in low power mode or
anything.

Signed-off-by: Anjandev Momi <anjan@momi.ca>