Disable bemenu hardware controll in lock / screenoff mode
If a bemenu is open while sxmo goes to lock then screenoff, hardware
buttons will continues to control it instead of giving access to the
sxmo state and so to unlock the device.

To fix it we simply enforce the default mode while locking /
screenoffing. We dont add additional behavior to go back to "menu" mode
on unlock if a menu is open. It means only the touch screen will be
able to control it but at least it prevent soft locking the device.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>