make sure touchscreen works if no deviceprofile is set
People have trouble writing a device profile when the touchscreen
doesnt work. The touchscreen doesnt work so you can't connect to wifi
to ssh in to write the device profile.

To solve this chicken and egg problem, this patch makes sure touchscreen
atleast works by introducing an "unknown" device. I
purposefully went against the linux guidelines for devicenames so that
the name will not conflict with an actual device.

Test procedure:

remove device profile for poco f1 in git repo and from
/usr/share/sxmo/default_hooks /usr/bin/

`doas make install` with this patch

Reboot and the touchscreen atleast still works.

Can someone please test this patch with a onebutton device? It doesnt
matter if the buttons dont work. Lets make sure the touchscreen atleast
works without a device profile.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>