alsamicismuted not implemented and led to 'Not implemented' appearing in status bar: write to stderr instead
I kept getting 'Not implemented' in my status bar, output from
sxmo_status.sh, which was caused by sxmo_audio.sh and a alsamicismuted()
outputting "Not implemented" to stdout, effectively obscuring most of
the tiny status bar. I redirected the output to stderr..

Ideally of course we should fix the function, but this will have to do
in the meantime.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>