Drop busybox aliases
As discussed with Aren over and over again, there is no easy, and fast enough
way to alias those in a smart way.

Cause this is causing issues on other distros than Alpine, we drop them
completly.

I think we should find a way to check those in the CI tests to
keep an eye on non busybox api usages contributors may try to use.

Signed-off-by: Stacy Harper <contact@stacyharper.net>
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>