Run statusbar asynchronously from lock state hooks
If the statusbar takes a while to run it will leave the system in an
inconsistent lock state, not waiting for the statusbar reduces the
amount of time we're in this state for.

This might be a slight performance improvement when switching through
multiple modes (e.g. screenoff to unlock) because the statusbar can take
a while to execute normally.

sxmo_daemons.sh is used to so that if the previous execution of the
statusbar is still running, it gets killed to avoid resource exhaustion.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>