Fix order of -maxdepth find argument
Gnu find outputs a warning if the maxdepth option appears after any
positional options, such as filtering files, because maxdepth also
affects options that come before it.

As a side note, maxdepth appears to be gnu extension that busybox picked
up, so we might want to be careful about using it.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>