Add a simple mode to bluetooth menu
I'm very proud of this bluetooth menu. It allow us to manage our
connection with ease. But most of the time the only thing I want to do
is to connect to devices. Having this device dedicated submenu as extra
step, forward and backward, just to connect, is really annoying.

This add a simple mode to this bluetooth menu. It is enabled by default.
If you select a device with this mode on, you'll connect/disconnect
automatically without entering the device dedicated submenu. For this
workflows to be confortable, I also added paired/connected icons on the
the device entries.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>