Don't sucks while being used with logind
This is not a logind support patch. Sxmo doesn't talk logind so using it
with sxmo is pretty much useless. You shouldn't use it.
At least, let's unbreak sxmo while being used with it.

The major problem is that sxmo XDG_RUNTIME_DIR fallback to its default
location when not present. This works for the graphical session, started
by tinydm, but this make ssh session and cronjobs to fails to grab sxmo
session.

We also have to set XDG_RUNTIME_DIR before probing sxmo to be sure to
use the same directory discovery rules.