Make proximity lock a real and dedicated sxmo state
This is a new approach to cleanup and fix issues we have with the
proximity lock. Discussed between Peter and me over irc.

We set "proximity[un]lock" in SXMO_STATE. This allow the input handler
to ignore power buttons while in this state. We then can't manually
change states anymore. It become possible to configure dedicated
hardware buttons trigger, or gesture trigger, to manage the current
call while in this state, by example.

The proximitylock also setup exclusive flock over the state lock
file (as other state switcher does). It prevent other state scripts to
trigger and to break the state consistency.

We stop the idle_locker to prevent 120s, and 8s timers to move the sxmo
state deeper.

When proximitylock stop, it execute back the initial state script,
to return to initial state.