MMS Massive Rewrite
The problem this MMS rewrite was supposed to address is the fact that
MMS writes the *contact names* and not the *phone numbers* to the log
files, whereas SMS writes the *phone numbers* and uses sed on
tailtextlog to conver these after the fact. SMS was doing things "the
right way" and MMS was doing things "the wrong way", so I wanted to
bring the two into harmony.

Note that this fix will *should* be fully backwards compatible too.

Along the way, I realized that we could split some things out. In
particular, I now have:

sxmo_hook_smslog.sh, which allows a user to change the look-and-feel of
how things get logged. (This might be not used very much.)

I also streamlined MMS code to make it more readable (since I couldn't
read bits of it after a year or so having written it).

I also implemented sxmo_contacts.sh --name-or-number (print the contact
name, and if there is no contact name print the number back at us, as
opposed to '???').  This cleaned up some code in other places.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>