start periodic_blink immediately after screenoff
There is already a single blink when the screen is turned off, and then
we wait for 5 seconds instead of the normal 2 before starting the
repeating blink. This replaces that with a single call to start the
repeating blink.

Also remove comments about dangling blinks, as they're always used now.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>