Avoid wakelock expiration if the check take much time
Never happend before but it is definitely possible that the 30s duration
could expire if the following trigger take longer than 20s. By example
in case of network issue.

The goal of this default duration is to avoid stucked wake
locks, even if this could theorically not happen anyway. So let's use a
big value to avoid causing troubles.