initial sxmo(7) man page
This is basically a conversion of USERGUIDE.md. Pushing this now for
comment on the structure.  The idea is to have each of our hooks as
its own man page (e.g., sxmo_hook_sms(1) and so on). And then we'd
also have "bigger" scripts have their own man pages, e.g.,
sxmo_files(1), sxmo_rtcwake(1), sxmo_notifications(1).

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>