major revision of sxmo(7) manpage
This revision changes things to refer to shell script for hooks not
manpages. I think manpages will just introduce problems. Instead, the
proposal is to have each hook include its own documentation in comments
at the start of it, and refer the viewer back to the section in the
sxmo(7) manpage where that hook is discussed.

This is unfinished but pushing it.

Signed-off-by: Peter John Hartman <peterjohnhartman@gmail.com>