sxmo_notification*.sh: Rework and cleanup
This is a complete rewrite, and group every scripts in a single
sxmo_notifs.sh one.

Also: Replace the watch file feature with a more generic notification group
one.

This feature make the user to miss notifications if the conversation file is
already open. It also cause problems when used with Syncthing, as this cause
notifications to be dropped when it synchronise files. It is also pointless
now that opening a notification to an opened conversation just focus the
corresponding terminal.

I think that dropping this feature completly is okay, make things simple,
and simplify the code

But in the same time:

Opening a notification should close all other ones for the same
conversations. So to cover this case, we add an optional group when
creating notifications. Discarding a notification now close all related
ones from the same group.