Cleanup udev rules
Main goal here is to remove obsoletes rules, and to drop our 666 chmods.
Those are serious security holes. Requiring the user to be in the "input"
group is already very weak. Let's not make this worse.

Brightnessctl embark its own rules, so we can drop our lights related
rules.

We move the touchscreen in a more common by-path folder. We are in the input
group, so lisgd have access to it.

I don't know about stylus related rules, I don't think they are
necessary. Input devices are in the "input" group by default. I don't
think those need to be 666...