Use a dedicated env variable to remove volume bind
Relying on SXMO_DEVICE_NAME isn't great cause some devices (ex: the
Pinebook Pro) got a device name that isn't "desktop" but are considered
by Sxmo with a desktop mode. Because of the symlink of the hook folder
to desktop.

I think a dedicated environment variable is good enough to cover this
case.