sxmo_rss.sh: Allow up to 2 http requests instead of only handling 1st
This changes to sxmo_rss.sh allows up to 2 http requests to be handled as some
rss feeds provide either a link, or a link and enclosure, an example of this is
PostmarketOS Podcast, with a link to the website and a link to the content in
the same file.

When the script detects more than one it breaks them into:
Link {1st http address}
Enclosure {2nd http address}

Otherwise everything happens as per the script originally.

This change line up with how sfeed_curses presents the links.

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Willow Barraco <contact@willowbarraco.fr>