apps: add Cozy
Cozy is an audiobook player:
<https://github.com/geigi/cozy>

it works well in portrait mode. in landscape mode the controls at the
bottom of the screen (play/pause/seek/speed) might not be visible on
small displays unless SXMO_SWAY_SCALE is reduced or some other mechanism
is used to avoid scaling the application.

Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>