state: rename down to idle and up to click
These match what the states are used for. These actions aren't the exact
inverse of each other, and with more states could even take different
routes through the available states.

For example if we were to dim the screen before entering lock, the idle
states would be: unlock, dim, lock, screenoff, and the click states
would be unlock, screenoff, lock. Logic would be needed to handle states
not in the click list.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>