sxmo_bluetoothmenu.sh: fix clean re-connection
Do not try to pair before connection. Let's stay consistent with how we
connect to new devices from the main loop. Also, do not trust by
default. The user have to do this themselves manually if they want to.

Add a --timeout to the scan, so that it stay scanning, even if not
in interactive mode.

Sleep 1 so that we don't spam too much the daemon.