sxmo_state.sh: prevent race condition while checking current state
Add a "get" command to abstract state checks, from inside of the
sxmo_state.sh script. This use the shared flock on this case.

This have to adapt the flocks management, so that the hooks can access
the "get" without staying idle, waiting for themselves to release the
exclusive flock.