Using xob for audio and brightness notifications in dwm
This patch replaces dunst with xob for audio and brightness levels.

Currently xob is only in alpine edge so this will be useful once it moves to
stable.

Xob behaves exactly the same as wob this the file for xob are just renamed from
wob. Alternative proposal is place a if statement in sxmo_wob.sh to start wob
or xob and then remove the $SXMO_WM check.

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Willow Barraco <contact@willowbarraco.fr>