alacritty: set behaviour to match other terminals
This patch fixes issues with using alacritty as your terminal emulator in sxmo.

Alacritty needs to know if you are trying to exrcute a specific program or $SHELL
as setting the title even with dynamic titles set in config will make the title
permanent

That affects calling sxmo_terminal.sh with power button as the title will stay
as the path to your shell.

To test for this made a check to see if the case is calling $* as shell path or
program

Also added to sxmo_hook_apps.sh

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>