proximitylock: do not restore initial state if the user changed it in between
To do this, we add store, restore, and flushstored to sxmo_state.sh.

store output an id that the caller have to give back to restore. This way,
multiple caller can store, and only the first one that restore actually
does so. This ensure some consistency to the end state workflow.

Any user click in between automatically flushstored, so that the end of
proximitylock will not restore to initial state.