hook_screenoff: blink purple only when suspend is enabled
When autosuspend is disabled, the purple LED blinks indefinitely, since
screenoff is the lowest idle state possible in that case. The purple LED
notification is meant to notify the user that that system is in screenoff, and
will likely suspend soon, thus for users who disable autosuspend, the blink
loses meaning. Disable it for this case.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>