Acme: add controls and selection options
As acme is a mouse heavy program it needs input from middle and right clicks.
As such this requires xdotool to simulate mouse button clicks.

To make it work I have bound volume up to right click and volume down to middle
click.

There is also options in context menu as backup if the device doesn't have volume
controls.

If the mouse 2 and 3 aren't clear they require you to reselect the spot in less
than 2 seconds for the mouse click to occur at that location.

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Aren Moynihan <aren@peacevolution.org>