contextmenu: detect terminal apps with case statement
This has the potential to cause odd breakages because it doesn't require
full word matches anymore. This can be mitigated by using *" word "*
instead of *"word"*, which I have done where I expect it to be a
problem.

Not needing to fork a bunch of printf / grep processes is a significant
performance improvement. I think this is well worth the risk of breaking
things.

Before:

$ hyperfine --warmup 4 -N sxmo_hook_contextmenu.sh
Benchmark 1: sxmo_hook_contextmenu.sh
  Time (mean ± σ):     421.0 ms ±   4.2 ms    [User: 282.9 ms, System: 233.7 ms]
  Range (min … max):   414.5 ms … 426.2 ms    10 runs

After:

$ hyperfine --warmup 4 -N sxmo_hook_contextmenu.sh
Benchmark 1: sxmo_hook_contextmenu.sh
  Time (mean ± σ):     151.3 ms ±   2.4 ms    [User: 118.8 ms, System: 105.1 ms]
  Range (min … max):   146.7 ms … 154.8 ms    19 runs

These tests were run on a PinePhone with a foot window focused using
bash as the shell.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>