statusbar: remove redundant symbols when modem has failed or unknown
This modifies the statusbar hook so that if there is a modem state that doesn't
match failed or unknown then show used techs and signal.

Currently if there isn't a sim in the modem there will be $icon_modem_failed, $icon_modem_notech, and $icon_modem_signal_0.

After this all that shows is $icon_modem_failed.

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>