sxmo_wm.sh: fix fringe cases in focused program
Found that the mpv pseudo-gui listed its WM_CLASS as `WM_CLASS(STRING) = "gl", "mpv"`

xorgfocusedwindow was only looking at the first field for the app: rename so it
was always set to gl. This patch changes the field to the second.

Testing on my machine hasn't found any issues as everything was detected
correctly on a few different windows.

Signed-off-by: hazardchem <hazardchem@disroot.org>
Signed-off-by: Aren Moynihan <aren@peacevolution.org>