sxmo_screenshot: workaround for slurp issues
We've already had a number of patches attempting to resolve this issue,
and an upstream PR to slurp pending for quite a while. I think we need a
solution in the meantime that would work even on the broken version of
slurp (in case distros are behind once a fix is released).

This version also copes with the fat-trembling-finger problem and will
simply take a screenshot of the whole screen if the selection is too
tiny to be reasonable (which may be an issue even in the fixed slurp).

Ref: https://lists.sr.ht/~mil/sxmo-devel/patches/49367
Ref: https://lists.sr.ht/~mil/sxmo-devel/patches/49719
Ref: https://lists.sr.ht/~mil/sxmo-devel/patches/44258
Ref: https://github.com/emersion/slurp/issues/140
Ref: https://github.com/emersion/slurp/pull/141
--
Changes in v2:
- select only one output (adapted from magdesign's patch earlier)
- better error feedback

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>