sms: fix newline EOL while sending messages
Some text editors, as kakoune, automatically add an ending newline while
saving buffers to files. This is because the convention on UNIX systems,
is that files ends with newline.

While sending sms using a file as resource, modem manager keep this. I
already discussed this with the maintainer:

https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/issues/830

So we basically have to do this ourselves.

The problems appeared when we switched to this new argument
--messaging-create-sms-with-text. It fixes our quoting issues, but it
introduce this problem, when the editor does the right thing.

To solve this, we create a temporary file, without this ending newline.
The finish script trap will clean in afterward.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>
Signed-off-by: Aren Moynihan <aren@peacevolution.org>