Makefile: cleanup handling of c compiler arguments
The c compiler seems to be rather picky about the order arguments are
defined in, I was having trouble building sxmo_status_led in some
conditions because -lm was in the wrong spot.

This is the order that these are supplied in the built-in rules
documentation of the gnu make manual.

While here, make cflags include any cflags passed by environment
variable, such as from a package build script.

Signed-off-by: Willow Barraco <contact@willowbarraco.fr>