launch screen lock if installed
This will launch peanutbutter (if it is installed) when the screen turns
off. It also omits the locked/screen-on stage then because it has less
added value and is in the way.

I'd recommend having users explicitly install peanutbutter themselves
(and read the manual) rather than have it pulled in automatically, as
users need to know the default unlock code or they get locked out of
there session. But this way no hooks need to be adapted if it's
installed, and it will out-of-the-box show the status icons on the
lockscreen.

Signed-off-by: Aren Moynihan <aren@peacevolution.org>