documentation: significant overhaul (man pages)
This is a significant overhaul of the Sxmo documentation. I'm building
upon the work Peter (wart) started, to provide proper man pages for Sxmo.
I think man pages fit our philosophy well, and ensure that all
essential documentation is always available offline on-device.

The user can simply do `man sxmo` and get the full manual (which was
called our 'user guide' in the old setup). I favour keeping this man
page as the sole source and automatically derive things for the
sxmo-docs repo (which is rather convoluted with multiple branches now).
Storing the documentation right-alongside sxmo-utils has the big
advantage of it describing (ideally) the exact version it is stored
alongside with. Submitted patches can do both implementation and
documentation in one go.

Things in this patch:

* Sections that were missing were added
* Some new man pages dedicated to specific scripts were added  (*.1.scd)
* Rewrote various parts to make the language better/clearer.
* Reordered things.
* Information that was out-of-date or incomplete has been updated (to
  the best of my knowledge)
* Applied a consistent style.
* Added HTML conversions in the Makefile (via groff and some hacky
  post-processing to insert decent hyperlinks, 'man' is a bit ancient)
* Added a 'Help' item in the main menu which opens the Sxmo man page in a terminal
* Updated the description of some hooks

This is not necessarily done yet (is documentation ever?), but it
provides a solid foundation I think. Some sections may need some further
work and everything should at least be checked thoroughly by others for
accuracy.